pyproject.toml
[tool.poetry]
name = "my-blog"
version = "0.1.0"
description = ""
authors = ["Semenov Yuriy <info@juriys.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
